tests/submission/Infraware/Forms/contents/Forms/textarea_checkvalidity.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>textarea_checkValidity</h3>
	</p>
	
	<hr>

	<div id="log"></div> 

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="" 
		  id="input_form">
		<p><textarea id='textarea_id'></textarea></p>
	</form>
	<script>

		var textarea = document.getElementById("textarea_id");

		try
		{
			var ret = textarea.checkValidity();

			test( function() {
				assert_equals(ret, true, "calling of checkValidity method is failed.");
			});
		}
		catch (e) {
			test( function() {
				assert_unreached("Error is raised.");
			});
		}

	</script>

 </body>
</html>
